refactor: constant 패키지 이동
